LP/bot/requirements.txt
matplotlib==3.2.1
python-telegram-bot==12.2.0
antlr4-python3-runtime==4.8